requirements.txt
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
SQLAlchemy>=2.0.0
python-dotenv>=1.0.1
rich>=13.7.0
